Move data utils from the script to package
